+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Unreleased
+-----------------------------
+* Use Tutor v1 plugin API
+
+
 Version 0.2.0 (2022-06-28)
 -----------------------------
 * Add `HASTEXO_ENABLE_SUSPENDER` and `HASTEXO_ENABLE_REAPER` to